+++ meteor-app/imports/api/assessments/server/publications.js
@@ -3,7 +3,10 @@ import Assessments from '../schema'
 import Services from '../services'
 import ServiceItems from '../../service-items/schema'
 import Logger from '../logger'
-import { JOB_STATUS_ALL, JOB_STATUS_COMPLETE } from '/imports/api/constants'
+import {
+  JOB_STATUS_ALL,
+  JOB_STATUS_COMPLETE,
+} from '/imports/api/constants'
 
 // Mainly to get updated changes reflected in job cart
 Meteor.publish('services.all', () => {
@@ -19,10 +22,13 @@ Meteor.publish('assessments.all', () => {
 })
 
 Meteor.publish('assessments.current', () => {
-  return Assessments.find({ status: { $in: JOB_STATUS_ALL } })
+  return Assessments.find(
+    { status: { $in: JOB_STATUS_ALL } },
+    { sort: { createdAt: -1 }, limit: 50 }
+  )
 })
 
-Meteor.publish('assessments.jobNo', jobNo => {
+Meteor.publish('assessments.jobNo', (jobNo) => {
   return Assessments.find({ jobNo })
 })
 
@@ -30,7 +36,7 @@ Meteor.publish('assessments.archive', () => {
   return Assessments.find({ status: { $in: JOB_STATUS_COMPLETE } })
 })
 
-Meteor.publish('logger.assessment', aId => {
+Meteor.publish('logger.assessment', (aId) => {
   if (aId) {
     return Logger.find({ aId })
   }